check consistency proof when fetching updates
This (very scrappy) code checks the consistency proof of the prison
register when fetching updates.  If the consistency proof fails, we
assume the data has been deleted and restarted, so we start again from
the beginning.

We don't actually check that the entries match the root hash, so a
register could still lie to us by sending entries in the RSF that
don't match the root hash.  This code defends against Murphy, but not
Macchiavelli.